move backend state from BTypes to components where it belongs
